Fix wrong function prototype

+++ include/functions.h
@@ -2707,7 +2707,7 @@ s8 func_8013D8DC(s8 arg0, GlobalContext* globalCtx);
 s8 func_8013D924(s16 arg0, GlobalContext* globalCtx);
 Actor* SubS_GetActorById(GlobalContext* globalCtx, Actor* actorListStart, u8 actorCategory, s16 actorId);
 s32 func_8013D9C8(GlobalContext* globalCtx, s16* arg1, s16* arg2, s32 arg3);
-u8 func_8013DB90(GlobalContext* globalCtx, Vec3f* arg1, f32 arg2);
+s32 func_8013DB90(GlobalContext* globalCtx, Vec3f* arg1, f32 arg2);
 s32 func_8013DC40(Path* arg0, s32 arg1, s32 arg2, Vec3f* arg3);
 void func_8013DCE0(GlobalContext* globalCtx, Vec3f* arg1, Actor* arg2, struct_8013DF3C_arg1* arg3, Path* arg4, s32 arg5, s32 arg6, s32 arg7, s32 arg8, u8 arg9);
 s32 func_8013DE04(GlobalContext* globalCtx, struct_8013DF3C_arg1* arg1, struct_8013DF3C_arg1_unk_func1 arg2, struct_8013DF3C_arg1_unk_func2 arg3, struct_8013DF3C_arg1_unk_func2 arg4, struct_8013DF3C_arg1_unk_func2 arg5);
